progress

+++ app.js
@@ -80,12 +80,23 @@ let quizQuestions = [
   </div>`
 ]
 
-let startPage = [
+let finishPage = [
+  `<h2>All done!</h2>
+  <form>
+    <p>
+      <label for="initial">Initials: </label>
+      <input type="text" id="initials">
+    </p>
+  </form>
+  <button type="submit" id="finish">Submit</buttton>
+  `,
   `
-  <h2>Coding Quiz Challenge</h2>
-  <p>Try to answer the following code-related questions within the time limit. Keep in mind that incorrect answers will
-  penalize your score/time by ten seconds!</p>
-  <button type='submit' id='startBtn'>Start Quiz</button>
+  <h2>Highscores:</h2>
+  <p id="scoreElem"></p>
+  <p>
+    <button id="clear">Clear</button>
+    <button id="goBack">Go Back</button>
+  </p>
   `
 ]
 
@@ -95,8 +106,13 @@ let startPage = [
 document.getElementById('startBtn').addEventListener('click', () => {
   
   let time = setInterval(() => {
-    count--
-    document.getElementById('timer').innerHTML = `Time: ${count}`
+    if (count > 0) {
+      count--
+      document.getElementById('timer').innerHTML = `Time: ${count}`
+    } else {
+      clearInterval(time)
+      document.getElementById('quiz').innerHTML = finishPage[0]
+    }
   }, 1000);
 
   // Start quiz
@@ -110,7 +126,7 @@ document.getElementById('startBtn').addEventListener('click', () => {
         document.getElementById('quiz').innerHTML = quizQuestions[questionPos]
 
         if(event.target.classList.contains('wrong')) {
-          count -= 10
+          count -= 20
         }
         if (event.target.classList.contains('correct')){
           score++
@@ -120,48 +136,46 @@ document.getElementById('startBtn').addEventListener('click', () => {
       else {
         document.getElementById('quiz').innerHTML = finishPage[0]
 
-        document.getElementById('finish').addEventListener('click', event => {
-          event.preventDefault()
+        document.getElementById('finish').addEventListener('click', () => {
 
-          console.log(score)
-          clearInterval(time)
-          document.getElementById('timer').textContent = `Time: 75`
-          localStorage.setItem('initials', document.getElementById('initials').value)
-          document.getElementById('quiz').innerHTML = finishPage[1]
-          console.log(localStorage.getItem('initials'))
+            document.getElementById('timer').textContent = `Time: 0`
 
-          document.getElementById('clear').addEventListener('click', () => {
-            localStorage.clear()
-            document.getElementById('scoreElem').textContent = ''
-          })
+            // let user = document.getElementById('initials').value
 
-          document.getElementById('goBack').addEventListener('click', () => {
-            document.getElementById('quiz').innerHTML = startPage[0]
-          })
-          })
-        }
+            localStorage.setItem('initials', document.getElementById('initials').value)
+            document.getElementById('scoreElem').append(document.getElementById('initials').value)
+            document.getElementById('initials').value = ''
+            
+            document.getElementById('quiz').innerHTML = finishPage[1]
+            
+            document.getElementById('clear').addEventListener('click', () => {
+              localStorage.clear()
+              document.getElementById('scoreElem').textContent = ''
+            })
+            
+            document.getElementById('goBack').addEventListener('click', () => {
+              location.href = './index.html'
+            })
+          
+        })
       }
+    }
   })
 })
 
-let finishPage = [
-  `<h2>All done!</h2>
-  <form>
-    <p>
-      <label for="initial">Initials: </label>
-      <input type="text" id="initials">
-    </p>
-    <button type="submit" id="finish">Submit</buttton>
-  </form>
-  `,
-  `
-  <h2>Highscores:</h2>
-  <p id="scoreElem">
-    ${localStorage.getItem('initials')} - ${score}
-  </p>
-  <p>
-    <button id="clear">Clear</button>
-    <button id="goBack">Go Back</button>
-  </p>
-  `
-]
+
+// when 'highscores' is clicked show highscores
+document.addEventListener('click', event => {
+  if(event.target.classList.contains('showHigh')) {
+    document.getElementById('quiz').innerHTML = finishPage[1]
+
+    document.getElementById('clear').addEventListener('click', () => {
+      localStorage.clear()
+      document.getElementById('scoreElem').textContent = ''
+    })
+
+    document.getElementById('goBack').addEventListener('click', () => {
+      location.href = './index.html'
+    })
+  }
+})
